DB Init: Add a script to ensure the docker container is available
It takes a few seconds for the docker container to start MySQL.
If we attempt to access the DB before MySQL starts, we'll hit
a bunch of 500s. This script is intended to terminate once
MySQL is started on the container, so we can effectively wait
before starting our app.